refactor: user 도메인 테스트 코드 수정
- spec.ts 파일을 spec 폴더 안에서 관리
- console.log 삭제
